# contracts/openapi.json
{
  "components": {
    "schemas": {
      "BatchLotResult": {
        "additionalProperties": false,
        "properties": {
          "actual_winners": {
            "description": "Who really won (demo)",
            "items": {
              "type": "string"
            },
            "title": "Actual Winners",
            "type": "array"
          },
          "invitations": {
            "items": {
              "$ref": "#/components/schemas/Invitation"
            },
            "title": "Invitations",
            "type": "array"
          },
          "lot": {
            "$ref": "#/components/schemas/LotListItem"
          },
          "winner_invited": {
            "description": "The real winner is among the invitations (demo)",
            "title": "Winner Invited",
            "type": "boolean"
          }
        },
        "required": [
          "lot",
          "invitations",
          "actual_winners",
          "winner_invited"
        ],
        "title": "BatchLotResult",
        "type": "object"
      },
      "BatchResponse": {
        "additionalProperties": false,
        "properties": {
          "day": {
            "description": "Publication date of the processed notices",
            "format": "date",
            "title": "Day",
            "type": "string"
          },
          "lots": {
            "items": {
              "$ref": "#/components/schemas/BatchLotResult"
            },
            "title": "Lots",
            "type": "array"
          },
          "timing_ms": {
            "title": "Timing Ms",
            "type": "number"
          },
          "total_invitations": {
            "title": "Total Invitations",
            "type": "integer"
          }
        },
        "required": [
          "day",
          "lots",
          "total_invitations",
          "timing_ms"
        ],
        "title": "BatchResponse",
        "type": "object"
      },
...
      "Invitation": {
        "additionalProperties": false,
        "description": "A supplier the robot would invite to the lot's procurement.",
        "properties": {
          "inn": {
            "title": "Inn",
            "type": "string"
          },
          "is_new": {
            "description": "From the SME registry, never bid before",
            "title": "Is New",
            "type": "boolean"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "role_display": {
            "title": "Role Display",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "inn",
          "name",
          "role_display",
          "score",
          "reason",
          "is_new"
        ],
        "title": "Invitation",
        "type": "object"
      },
      "LotCard": {
        "additionalProperties": false,
        "properties": {
          "actual_winners": {
            "description": "INNs of the real winners, for lots from the data (demo)",
            "items": {
              "type": "string"
            },
            "title": "Actual Winners",
            "type": "array"
          },
          "channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "customer_inn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Inn"
          },
          "is_smp": {
            "title": "Is Smp",
            "type": "boolean"
          },
          "items": {
            "items": {
              "type": "string"
            },
            "title": "Items",
            "type": "array"
          },
          "lot_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "None for a new lot",
            "title": "Lot Id"
          },
          "okpd2_codes": {
            "items": {
              "type": "string"
            },
            "title": "Okpd2 Codes",
            "type": "array"
          },
          "procedure_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Procedure Name"
          },
          "publish_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publish Date"
          },
          "start_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Price"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          }
        },
        "required": [
          "lot_id",
          "publish_date",
          "subject",
          "procedure_name",
          "start_price",
          "okpd2_codes",
          "items",
          "customer_inn",
          "channel",
          "is_smp",
          "actual_winners"
        ],
        "title": "LotCard",
        "type": "object"
      },
      "LotListItem": {
        "additionalProperties": false,
        "properties": {
          "channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "customer_inn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Inn"
          },
          "lot_id": {
            "title": "Lot Id",
            "type": "integer"
          },
          "publish_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publish Date"
          },
          "start_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Price"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          }
        },
        "required": [
          "lot_id",
          "publish_date",
          "subject",
          "start_price",
          "channel",
          "customer_inn"
        ],
        "title": "LotListItem",
        "type": "object"
      },
...
      "ValidationErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ValidationErrorResponse",
        "type": "object"
      },
      "ValidationIssue": {
        "additionalProperties": false,
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Loc",
            "type": "array"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationIssue",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Recommendation service and Explainable AI (XAI) API for counterparty matching",
    "title": "Roseltorg • Intelligent Counterparty Selection Service AIS GZ",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/batch/simulate": {
      "post": {
        "description": "The background robot for one day: newly published notices, suppliers to invite for each\n(known ones with the reason, plus new companies from the SME registry).",
        "operationId": "simulate_batch",
        "parameters": [
          {
            "description": "Notices of this day; default: the last",
            "in": "query",
            "name": "day",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Notices of this day; default: the last",
              "title": "Day"
            }
          },
          {
            "in": "query",
            "name": "max_lots",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Max Lots",
              "type": "integer"
            }
          },
          {
            "description": "Known suppliers to invite",
            "in": "query",
            "name": "per_lot",
            "required": false,
            "schema": {
              "default": 5,
              "description": "Known suppliers to invite",
              "maximum": 20,
              "minimum": 1,
              "title": "Per Lot",
              "type": "integer"
            }
          },
          {
            "description": "New registry companies",
            "in": "query",
            "name": "new_per_lot",
            "required": false,
            "schema": {
              "default": 2,
              "description": "New registry companies",
              "maximum": 10,
              "minimum": 0,
              "title": "New Per Lot",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Simulate Batch Recommendations",
        "tags": [
          "Batch"
        ]
      }
    },
...
    "/api/v1/lots/search": {
      "get": {
        "description": "Find lots by subject text, lot id or registry number; newest first.",
        "operationId": "search_lots",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "default": "",
              "title": "Query",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LotListItem"
                  },
                  "title": "Response Search Lots",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Lots",
        "tags": [
          "Lots"
        ]
      }
    },
    "/api/v1/lots/{lot_id}": {
      "get": {
        "description": "A lot with its items, OKPD2 codes and the real winners.",
        "operationId": "get_lot",
        "parameters": [
          {
            "in": "path",
            "name": "lot_id",
            "required": true,
            "schema": {
              "title": "Lot Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LotCard"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Lot",
        "tags": [
          "Lots"
        ]
      }
    },
...
  }
}